+++ src/test/java/edu/hw1/Task8Test.java
@@ -1,0 +1,44 @@
+package edu.hw1;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
+import java.util.stream.Stream;
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+
+public class Task8Test {
+    private static Stream<Arguments> knightsTestParameters() {
+        return Stream.of(
+            Arguments.of(new int[][] {
+                {0, 0, 0, 1, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 1, 0, 0, 0, 1, 0, 0},
+                {0, 0, 0, 0, 1, 0, 1, 0},
+                {0, 1, 0, 0, 0, 1, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 1, 0, 0, 0, 0, 0, 1},
+                {0, 0, 0, 0, 1, 0, 0, 0}
+            }, true),
+            Arguments.of(new int[][] {
+                {1, 0, 1, 0, 1, 0, 1, 0},
+                {0, 1, 0, 1, 0, 1, 0, 1},
+                {0, 0, 0, 0, 1, 0, 1, 0},
+                {0, 0, 1, 0, 0, 1, 0, 1},
+                {1, 0, 0, 0, 1, 0, 1, 0},
+                {0, 0, 0, 0, 0, 1, 0, 1},
+                {1, 0, 0, 0, 1, 0, 1, 0},
+                {0, 0, 0, 1, 0, 1, 0, 1}
+            }, false)
+        );
+    }
+
+    @DisplayName("Кони на доске")
+    @ParameterizedTest
+    @MethodSource("knightsTestParameters")
+    void knightsTest(int[][] deck, boolean expected) {
+        boolean result = Task8.knightBoardCapture(deck);
+
+        assertThat(result).isEqualTo(expected);
+    }
+}
